order-generator/requirements.txt
faker
Flask
boto3
pandas
google-cloud-storage